refactor(prompt): reorganize sources
