+++ @commitlint/prompt/package.json
@@ -1,16 +1,27 @@
 {
   "name": "@commitlint/prompt",
   "version": "3.0.0",
-  "description": "commit prompt using .commitlintrc",
-  "bin": {
-    "commit-prompt": "./lib/index.js"
-  },
+  "description": "commitizen prompt using .commitlintrc",
+  "main": "./lib/index.js",
   "scripts": {
     "build": "cross-env NODE_ENV=production babel src --out-dir lib",
     "clean": "rimraf lib",
+    "commit": "git-cz",
     "pretest": "dep-check",
-    "test": "exit 0",
+    "test": "ava",
     "prepublish": "npm run build"
+  },
+  "ava": {
+    "babel": "inherit",
+    "require": [
+      "babel-register"
+    ],
+    "files": [
+      "src/**/*.test.js"
+    ],
+    "sources": [
+      "src/**/*.js"
+    ]
   },
   "babel": {
     "env": {
@@ -18,7 +29,6 @@
         "sourceMaps": "inline",
         "plugins": [
           "add-module-exports",
-          "istanbul",
           [
             "transform-runtime",
             {
@@ -56,6 +66,11 @@
       ]
     ]
   },
+  "config": {
+    "commitizen": {
+      "path": "./@commitlint/prompt"
+    }
+  },
   "xo": false,
   "repository": {
     "type": "git",
@@ -76,19 +91,26 @@
   "homepage": "https://github.com/marionebl/commitlint#readme",
   "devDependencies": {
     "@commitlint/utils": "^3.0.0",
+    "ava": "^0.20.0",
     "babel-cli": "^6.24.1",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-preset-env": "1.6.0",
     "babel-preset-stage-0": "^6.24.1",
-    "cross-env": "5.0.1"
+    "babel-register": "^6.24.1",
+    "commitizen": "^2.9.6",
+    "cross-env": "5.0.1",
+    "throat": "^4.1.0"
   },
   "dependencies": {
     "@commitlint/core": "^3.0.0",
     "babel-polyfill": "^6.23.0",
     "babel-runtime": "^6.23.0",
     "chalk": "1.1.1",
-    "lodash": "4.5.1",
+    "execa": "^0.7.0",
+    "lodash": "^4.17.4",
+    "meow": "^3.7.0",
+    "throat": "^4.1.0",
     "vorpal": "1.10.0"
   }
 }
