+++ @commitlint/prompt/src/index.js
@@ -1,5 +1,6 @@
 import 'babel-polyfill';  // eslint-disable-line import/no-unassigned-import
-import getInput from './get-input';
+import vorpal from 'vorpal';
+import input from './input';
 
 /**
  * Entry point for commitizen
@@ -7,12 +8,7 @@ import getInput from './get-input';
  * @param {function} commit callback to execute with complete commit message
  * @return {string} genersated commit message
  */
-export const prompter = async ({prompt}, commit) => { // eslint-disable-line import/prefer-default-export
-	getInput(prompt)
-		.then(result => commit(result))
-		.catch(err => {
-			setTimeout(() => {
-				throw err;
-			}, 0);
-		});
+export const prompter = async (_, commit) => { // eslint-disable-line import/prefer-default-export
+	const message = await input(vorpal);
+	commit(message);
 };
