+++ @commitlint/prompt/src/library/rule-is-active.js
@@ -1,0 +1,9 @@
+/**
+ * Check if a rule definition is active
+ * @param {object} rule to check
+ * @return {boolean} if the rule definition is active
+ */
+export default function ruleIsActive(rule) {
+	const [, [severity]] = rule;
+	return severity > 0;
+}
